feat(frontend): filter menu options on Settings page (#1900)
# Motivation

The menu option "Your addresses" does not work on the "Settings" page.
Moreover, display a "Settings" link in the menu on the "Settings" page
does no add much value. Therefore this PR filters the options displayed
in the menu when the user visit the "Settings" page.

# Changes

- Refactor options related to the wallet from the menu in a new
component `MenuWallet`
- Do not display wallet options and "Settings" option in the menu when
user visit the "Settings" page

# Screenshot

This option does not work:

<img width="1536" alt="Capture d’écran 2024-07-29 à 08 58 01"
src="https://github.com/user-attachments/assets/7ed0c8b8-f6d3-4e96-a641-daa1cebbef13">
