+++ src/frontend/src/lib/components/core/Menu.svelte
@@ -6,24 +6,15 @@
 	import { OISY_REPO_URL } from '$lib/constants/oisy.constants';
 	import IconWallet from '$lib/components/icons/IconWallet.svelte';
 	import IconChevronDown from '$lib/components/icons/IconChevronDown.svelte';
-	import {
-		networkEthereum,
-		networkICP,
-		networkId,
-		pseudoNetworkChainFusion
-	} from '$lib/derived/network.derived';
-	import { networkParam } from '$lib/utils/nav.utils';
-	import EthWalletAddress from '$eth/components/core/EthWalletAddress.svelte';
-	import IcWalletAddress from '$icp/components/core/IcWalletAddress.svelte';
+	import { networkId } from '$lib/derived/network.derived';
+	import { isRouteSettings, networkParam } from '$lib/utils/nav.utils';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
 	import IconSettings from '$lib/components/icons/IconSettings.svelte';
 	import IconGitHub from '$lib/components/icons/IconGitHub.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
 	import ButtonMenu from '$lib/components/ui/ButtonMenu.svelte';
-	import WalletAddresses from '$lib/components/core/WalletAddresses.svelte';
-	import IconWalletConnect from '$lib/components/icons/IconWalletConnect.svelte';
-	import { modalStore } from '$lib/stores/modal.store';
-	import { walletConnectPaired } from '$eth/stores/wallet-connect.store';
+	import MenuWallet from '$lib/components/core/MenuWallet.svelte';
+	import { page } from '$app/stores';
 
 	let visible = false;
 	let button: HTMLButtonElement | undefined;
@@ -33,10 +24,11 @@
 		await goto(`/settings?${networkParam($networkId)}`);
 	};
 
-	const openWalletConnectAuth = () => {
-		visible = false;
-		modalStore.openWalletConnectAuth();
-	};
+	let settingsRoute = false;
+	$: settingsRoute = isRouteSettings($page);
+
+	let walletOptions = true;
+	$: walletOptions = !settingsRoute;
 </script>
 
 <button
@@ -50,31 +42,16 @@
 
 <Popover bind:visible anchor={button} direction="rtl">
 	<div class="flex flex-col gap-4">
-		{#if $networkICP}
-			<IcWalletAddress />
-		{:else if $networkEthereum}
-			<EthWalletAddress />
-		{:else if $pseudoNetworkChainFusion}
-			<WalletAddresses on:icReceiveTriggered={() => (visible = false)} />
+		{#if walletOptions}
+			<MenuWallet on:icMenuClick={() => (visible = false)} />
 		{/if}
 
-		<Hr />
-
-		<ButtonMenu
-			ariaLabel={$i18n.wallet_connect.text.name}
-			on:click={openWalletConnectAuth}
-			disabled={$walletConnectPaired}
-		>
-			<IconWalletConnect />
-			{$i18n.wallet_connect.text.name}
-		</ButtonMenu>
-
-		<Hr />
-
-		<ButtonMenu ariaLabel={$i18n.navigation.alt.more_settings} on:click={gotoSettings}>
-			<IconSettings />
-			{$i18n.settings.text.title}
-		</ButtonMenu>
+		{#if !settingsRoute}
+			<ButtonMenu ariaLabel={$i18n.navigation.alt.more_settings} on:click={gotoSettings}>
+				<IconSettings />
+				{$i18n.settings.text.title}
+			</ButtonMenu>
+		{/if}
 
 		<ExternalLink
 			href="https://identity.ic0.app"
